refactor: MediaType enum 객체로 변환
