+++ mvc/src/main/java/com/interface21/web/bind/annotation/RequestMethod.java
@@ -1,5 +1,9 @@
 package com.interface21.web.bind.annotation;
 
 public enum RequestMethod {
-    GET, HEAD, POST, PUT, PATCH, DELETE, OPTIONS, TRACE
+    GET, HEAD, POST, PUT, PATCH, DELETE, OPTIONS, TRACE;
+
+    public static RequestMethod from(String method) {
+        return RequestMethod.valueOf(method);
+    }
 }
