+++ mvc/src/main/java/com/interface21/web/http/MediaType.java
@@ -1,6 +1,20 @@
 package com.interface21.web.http;
 
-public class MediaType {
-    public static final String APPLICATION_JSON_UTF8_VALUE = "application/json;charset=UTF-8";
-    public static final String FORM_URL_ENCODED = "application/x-www-form-urlencoded";
+public enum MediaType {
+    APPLICATION_JSON_UTF8_VALUE("application/json;charset=UTF-8"),
+    FORM_URL_ENCODED("application/x-www-form-urlencoded");
+
+    private final String value;
+
+    MediaType(String value) {
+        this.value = value;
+    }
+
+    public static MediaType from(String value) {
+        return MediaType.valueOf(value);
+    }
+
+    public String getValue() {
+        return value;
+    }
 }
