+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/view/JsonView.java
@@ -14,7 +14,7 @@ public class JsonView implements View {
 
     @Override
     public void render(final Map<String, ?> model, final HttpServletRequest request, HttpServletResponse response) throws Exception {
-        response.setContentType(MediaType.APPLICATION_JSON_UTF8_VALUE);
+        response.setContentType(MediaType.APPLICATION_JSON_UTF8_VALUE.getValue());
         response.getWriter().write(mapper.writeValueAsString(model));
     }
 }
