+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/handler/AnnotationHandlerMappingTest.java
@@ -123,7 +123,7 @@ class AnnotationHandlerMappingTest {
         final var response = mock(HttpServletResponse.class);
 
         when(request.getRequestURI()).thenReturn("/users-json");
-        when(request.getHeader("Content-Type")).thenReturn(MediaType.APPLICATION_JSON_UTF8_VALUE);
+        when(request.getHeader("Content-Type")).thenReturn(MediaType.APPLICATION_JSON_UTF8_VALUE.getValue());
         when(request.getMethod()).thenReturn("POST");
 
         String body = "{\"userId\":\"gugu\",\"password\":\"123\",\"age\":\"19\"}";
@@ -169,7 +169,7 @@ class AnnotationHandlerMappingTest {
         final var response = mock(HttpServletResponse.class);
 
         when(request.getRequestURI()).thenReturn("/users-model");
-        when(request.getHeader("Content-Type")).thenReturn(MediaType.FORM_URL_ENCODED);
+        when(request.getHeader("Content-Type")).thenReturn(MediaType.FORM_URL_ENCODED.getValue());
         when(request.getMethod()).thenReturn("POST");
 
         String body = "userId=gugu&password=123&age=19";
